Latest changes to UI and functionality

+++ src/components/Login.js
@@ -3,8 +3,6 @@ import Avatar from "@mui/material/Avatar";
 import Button from "@mui/material/Button";
 import CssBaseline from "@mui/material/CssBaseline";
 import TextField from "@mui/material/TextField";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import Checkbox from "@mui/material/Checkbox";
 import Link from "@mui/material/Link";
 import Grid from "@mui/material/Grid";
 import Box from "@mui/material/Box";
@@ -13,6 +11,7 @@ import Typography from "@mui/material/Typography";
 import Container from "@mui/material/Container";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import axios from "axios";
+import Swal from "sweetalert2";
 
 function Copyright(props) {
   return (
@@ -49,12 +48,20 @@ export default function Login(props) {
         password: password,
       })
       .then(function (response) {
-        localStorage.setItem(
-          "authToken",
-          JSON.stringify(response.data.authorization)
-        );
-
-        window.location.href = "/";
+        if (response.data.authorization === undefined) {
+          Swal.fire({
+            title: "Error!",
+            text: "Invalid Email Or Password!",
+            icon: "error",
+            confirmButtonText: "Close",
+          });
+        } else {
+          localStorage.setItem(
+            "authToken",
+            JSON.stringify(response.data.authorization)
+          );
+          window.location.href = "/videos";
+        }
       })
       .catch(function (error) {
         console.log(error);
